In merge_rawnc, move 'badfiles' into the same for-loop as 'goodfiles'
This change is for consistency with a previous fix to the 'goodfiles'
variable, which was moved into a different for-loop in order to
properly handle the situation where we have .pld files but no .gld
files. The 'badfiles' variable is used for logging, and will now
log all bad files at the end of the processing loop.