Update depth and time attributes for IOOS and fix tests
Updated the time calendar in ncprocess.py, add depth
attributes in utils.py, change processing script to use
the new dbd reader code in test_pyglider.py (with option
to use the old processing), and reran the 3 netcdf files
that the tests compare against so that those files will
include the updated attributes. The size of the slocum
netcdf file is smaller than the old file due to moving
to the new dbdreader processing, which uses sci_water_temp
as the time base, instead of sci_m_present_time. This
change results in far fewer nans in the final dataset.